requirements.txt
streamlit==1.32.0
streamlit-authenticator==0.3.1
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.30
langchain-google-genai==0.0.11
google-generativeai==0.4.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
pyyaml==6.0.1
pypdf2==3.0.1